added ConcreteUserController and CurrentUserController(edited methods: getCurrentUser, editNameEmail, editPassword). Many checks, documentation, formatting and related changes were added, a new Organization model was added, the Serializable problem was solved, the DTO structures were changed

+++ src/main/java/com/dazito/oauthexample/model/Organization.java
@@ -14,12 +14,9 @@ import java.util.List;
 @NoArgsConstructor
 public class Organization {
 
-    //TODO: dto
-    //TODO: converter
-
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @Column(name = "org_name")
     private String organizationName;
